requirements.txt
fastapi==0.116.1
openai==1.97.0
python-dotenv==1.1.1
pydantic_settings==2.10.1
SQLAlchemy==2.0.41
Pillow==11.3.0
pdf2image==1.17.0
uvicorn==0.35.0